+++ src/main.rs
@@ -15,6 +15,7 @@
  */
 
 pub mod api;
+pub mod tool;
 mod command;
 pub mod config;
 mod history_file;
@@ -91,6 +92,13 @@ fn main() -> io::Result<()> {
                 app_config.user_config.max_tokens,
             );
             rebuild_chat_client = false;
+        }
+
+        // TODO: This shouldn't be printed on every iteration and model information should be fetched once
+        if &app_config.current_profile.provider == "ollama" {
+            if chat_client.model_supports_tools() {
+                println!("Model supports tools");
+            }
         }
 
         // Read the context file if provided